requirements.txt
Flask==3.1.0
waitress==3.0.2
requests
numpy
python-dotenv